ref: Remove unnecessary component
